Update Inventory syncing to ingnore unused and irrelevant frames

+++ config/broadsign.php
@@ -14,14 +14,14 @@ return [
         "key" => env("BROADSIGN_API_KEY"),
     ],
 
-    "domain-id" => env("BROADSIGN_DOMAIN_ID"),
-    "customer-id" => env("BROADSIGN_CUSTOMER_ID"),
+    "domain-id" => (int)env("BROADSIGN_DOMAIN_ID"),
+    "customer-id" => (int)env("BROADSIGN_CUSTOMER_ID"),
 
     "default-campaign-length" => 10, // years
 
-    "category-separation-id" => env("BROADSIGN_CATEGORY_SEPARATION_ID"),
+    "category-separation-id" => (int)env("BROADSIGN_CATEGORY_SEPARATION_ID"),
 
-    "advertising-criteria" => env("BROADSIGN_ADVERTISING_CRITERIA_ID"),
-    "left-frame-criteria" => env("BROADSIGN_LEFT_FRAME_CRITERIA_ID"),
-    "right-frame-criteria" => env("BROADSIGN_RIGHT_FRAME_CRITERIA_ID"),
+    "advertising-criteria" => (int)env("BROADSIGN_ADVERTISING_CRITERIA_ID"),
+    "left-frame-criteria" => (int)env("BROADSIGN_LEFT_FRAME_CRITERIA_ID"),
+    "right-frame-criteria" => (int)env("BROADSIGN_RIGHT_FRAME_CRITERIA_ID"),
 ];
